refactor[#523] : features/auth/api 컨벤션에 맞춰 타입 리팩토링
